PFP-8907 endret opprettsmanuellbehandling for å hente aktørId fra fpsak
