+++ domenetjenester/behandling/src/main/java/no/nav/foreldrepenger/tilbakekreving/behandling/impl/BehandlingUtil.java
@@ -1,0 +1,29 @@
+package no.nav.foreldrepenger.tilbakekreving.behandling.impl;
+
+import java.time.DayOfWeek;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.Period;
+
+import no.nav.vedtak.util.FPDateUtil;
+
+class BehandlingUtil {
+
+    BehandlingUtil(){
+        // for CDI proxy
+    }
+
+    static LocalDateTime bestemFristForBehandlingVent(LocalDate frist, Period defaultVentefrist) {
+        return frist != null
+            ? LocalDateTime.of(frist, FPDateUtil.nå().toLocalTime())
+            : FPDateUtil.nå().plus(defaultVentefrist);
+    }
+
+    static boolean sjekkAvvikHvisSisteDagIHelgen(LocalDate sisteDag, int antallDager) {
+        if (antallDager == 3 && sisteDag.getDayOfWeek() == DayOfWeek.FRIDAY) {
+            return false;
+        }
+        return antallDager != 2 || sisteDag.getDayOfWeek() != DayOfWeek.SATURDAY;
+    }
+
+}
