+++ web/src/test/java/no/nav/foreldrepenger/tilbakekreving/web/app/tjenester/behandling/aksjonspunkt/AksjonspunktRestTjenesteTest.java
@@ -24,6 +24,7 @@ import org.junit.rules.ExpectedException;
 
 import no.nav.foreldrepenger.tilbakekreving.behandling.BehandlingTjeneste;
 import no.nav.foreldrepenger.tilbakekreving.behandling.impl.BehandlingTjenesteImpl;
+import no.nav.foreldrepenger.tilbakekreving.behandling.impl.FeilutbetalingTjeneste;
 import no.nav.foreldrepenger.tilbakekreving.behandlingskontroll.BehandlingskontrollAsynkTjeneste;
 import no.nav.foreldrepenger.tilbakekreving.behandlingskontroll.BehandlingskontrollProvider;
 import no.nav.foreldrepenger.tilbakekreving.behandlingskontroll.BehandlingskontrollTjeneste;
@@ -81,6 +82,7 @@ public class AksjonspunktRestTjenesteTest {
                 mock(SimuleringIntegrasjonTjeneste.class),
                 mock(FagsakTjeneste.class),
                 mock(HistorikkinnslagTjeneste.class),
+                mock(FeilutbetalingTjeneste.class),
                 mock(FpsakKlient.class),
                 DEFAULT_PERIOD);
